类别的增删改

+++ CamuseHome/Models/Picture.cs
@@ -8,11 +8,11 @@ namespace CamuseHome.Models
 {
     public class Picture
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Path { get; set; }
         public string Remark { get; set; }
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
         public virtual Product Product { get; set; }
     }
 }
